ai improvement
seperate ai and the engine.py
